Fix documentation, change default_dir to contrib
Fixed dependencies in Makefile.PL.

Updated documented of dependencies in README.md.

The default_dir key in the config file was renamed to contrib. It now
defaults to the share directory distributed along with the code (and
installed via cpan).

Fixed the location of the config file. By default, it is read from the
location where the script is, which makes no sense when the
application is installed via cpan. It is now read from the current
directory.