Add a Regenerate button for text description template
This is the template used for the nomap examples, and nomap is used by
the examples on the Help page. Without this change, there's no way to
"reload" since the seed is passed along such that the Markdown button
generates the same text. The only way to regenerate the output would
be to go back, shift-reload to regenerate the invisible seed, and then
submit the form again. The Regenerate button now has a different seed
that gets used instead.